power-api/Code/requirements.txt
Flask==1.0.4
requests==2.23.0
numpy==1.17.3
boto3==1.13.26
pandas==0.25.2
aiohttp==3.6.2
kafka-python==2.0.1
python_dateutil==2.8.1
Gunicorn==20.0.4
